Added motorcycles and tasks

+++ MotorcycleMaitenance/data/CDImporter.swift
@@ -42,8 +42,8 @@ struct JsonMotorcycleMaintenance : Codable {
 }
 
 struct JsonMotorcycleTypeTask : Codable {
-    let mileageInterval: Int16
-    let timeInterval: Int16
+    let mileageInterval: Int32
+    let timeInterval: Int32
     let taskId: String
     let remarks: String?
 }
@@ -57,7 +57,7 @@ struct JsonMotorcycleMaintenanceWithTasks : Codable {
     struct JsonMotorcycleMaintenanceTask : Codable {
         let taskId: String
         let completionDate: Date
-        let mileage: Int16
+        let mileage: Int32
         let remarks: String
     }
